+++ io_revolt/ncp_out.py
@@ -50,6 +50,7 @@ def export_file(filepath, scene):
                 not obj.revolt.is_cube and
                 not obj.revolt.is_bcube and
                 not obj.revolt.is_bbox and
+                #not obj.revolt.is_instance and
                 not obj.revolt.ignore_ncp
             )
             if conditions: