+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog 
 
+## 2018-06-
+
+Version [`rva_18.06`](https://github.com/Yethiel/re-volt-addon/releases/tag/rva_18.06)
+
+**Topics**: Bugfixes, Tools
+
+- **Fixes**
+    - Batch bake should not crash anymore
+    - Import and export of .fin properties
+- **Additions**
+    - Batch Bake section in the Light and Shadow tool panel
+    - RVGL Launcher in the Helpers panel. The RVGL path (folder) has to be set in the add-on settings.
+- **Modifications**
+    - 
+
 ## 2018-06-19
 
 Version [`rva_18.0619`](https://github.com/Yethiel/re-volt-addon/releases/tag/rva_18.0619)